+++ Rebus.GoogleCloudStorage.Tests/DataBus/FakeRebusTime.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using Rebus.Time;
+
+namespace Rebus.GoogleCloudStorage.Tests.DataBus
+{
+    internal class FakeRebusTime : IRebusTime
+    {
+        DateTimeOffset? _fakeNow;
+
+        public DateTimeOffset Now => _fakeNow ?? DateTimeOffset.Now;
+
+        public void SetNow(DateTimeOffset time)
+        {
+            _fakeNow = time;
+        }
+    }
+}