+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## 1.0.0
 * Initial version
 
+## 1.0.2
+* Updated Rebus library dependencies
+
+## 1.0.3
+* Added support for the missing IDataBusStorageManagement API's to be able to query and delete existing objects on the data bus.
+
+
 ---
 
 [kendallb]: https://github.com/kendallb